fix commit: frontend修正: src/features/homeをfeatures/list/に命名を修正 | src/app/page.tsxのTableコンポーネント内容をfeatures/list/list.tsxに移動しimportで利用するように修正
